# Conflicts:
#	src/main/resources/mzmineversion.properties

+++ src/main/java/io/github/mzmine/modules/io/projectsave/ProjectSavingTask.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2021 The MZmine Development Team
+ * Copyright 2006-2022 The MZmine Development Team
  *
  * This file is part of MZmine.
  *
@@ -29,6 +29,7 @@ import io.github.mzmine.taskcontrol.AbstractTask;
 import io.github.mzmine.taskcontrol.TaskStatus;
 import io.github.mzmine.util.ExceptionUtils;
 import io.github.mzmine.util.StreamCopy;
+import io.github.mzmine.util.files.FileAndPathUtil;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -63,7 +64,8 @@ public class ProjectSavingTask extends AbstractTask {
   private UserParameterSaveHandler userParameterSaveHandler;
 
   private final int totalSaveItems;
-  private int currentStage, finishedSaveItems = 0;
+  private final int finishedSaveItems = 0;
+  private int currentStage;
   private String currentSavedObjectName;
 
   // This hashtable maps raw data files to their ID within the saved project
@@ -152,6 +154,14 @@ public class ProjectSavingTask extends AbstractTask {
       // Prepare a temporary ZIP file. We create this file in the same
       // directory as the final saveFile to avoid moving between
       // filesystems in the last stage (renameTo)
+
+      // checking if directory exists first
+      if (!FileAndPathUtil.createDirectory(saveFile.getParentFile())) {
+        setErrorMessage("Could not create directories for file " + saveFile + " for writing.");
+        setStatus(TaskStatus.ERROR);
+        return;
+      }
+
       File tempFile = File.createTempFile(saveFile.getName(), ".tmp", saveFile.getParentFile());
       tempFile.deleteOnExit();
